requirements.txt
certifi==2017.4.17
chardet==3.0.4
idna==2.5
oauthlib==2.0.2
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.1
requests-oauthlib==0.8.0
six==1.11.0
TwitterAPI==2.4.5
urllib3==1.21.1
